+++ CHANGELOG.md
@@ -14,3 +14,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - support displaying video
 - add dialog with direction details and map to geo-direction
 - create a calendar event
+- create a classified item